+++ custom_components/jablotron_futura/sensor.py
@@ -93,7 +93,7 @@ class FuturaSummarySensorEntity(FuturaEntity, SensorEntity):
 
     @property
     def unique_id(self) -> str:
-        return self.entity_description.key
+        return f"{self._central_unit.serial_no}_{self.entity_description.key}"
 
     @property
     def available(self) -> bool:
@@ -137,7 +137,7 @@ class FuturaPeripherySensorEntity(FuturaEntity, SensorEntity):
 
     @property
     def unique_id(self) -> str:
-        return self.entity_description.key
+        return f"{self._central_unit.serial_no}_{self.entity_description.key}"
 
     @property
     def native_value(self) -> StateType | date | datetime: